# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: QuimiPort API
  version: 1.0.0
  description: |
    API backend do **QuimiPort**, sistema de gestão de cargas químicas inspirado nas operações do Porto de Santos.

    ## Principais regras de negócio
    - Produto químico não pode ser cadastrado sem **nome**, sem **classe de risco** ou com **status inválido**.
    - Produto **inativo** não pode ser usado em novas cargas.
    - Carga exige **produto associado**, **quantidade > 0** e **responsável técnico**.
    - Carga só é **liberada** com a documentação obrigatória (**FISPQ** e **FICHA_EMERGENCIA**) dentro da validade.
    - Toda mudança de status passa pela máquina de estados abaixo; o que não está nela é proibido.

    | De | Para (permitido) |
    |---|---|
    | REGISTRADA | EM_ANALISE, CANCELADA |
    | EM_ANALISE | EM_INSPECAO, CANCELADA |
    | EM_INSPECAO | LIBERADA, BLOQUEADA, CANCELADA |
    | LIBERADA | EM_MOVIMENTACAO, CANCELADA |
    | BLOQUEADA | CANCELADA |
    | EM_MOVIMENTACAO | FINALIZADA |
    | FINALIZADA | — (terminal) |
    | CANCELADA | — (terminal) |

    ## Formato de erro
    Todos os erros seguem `{ "erro": { "codigo", "mensagem", "detalhes" } }`.

    | HTTP | codigo | Quando |
    |---|---|---|
    | 400 | VALIDATION_ERROR | Campo ausente/inválido |
    | 404 | NOT_FOUND | Recurso inexistente |
    | 409 | CONFLICT | Código de carga duplicado |
...
    | 500 | INTERNAL_ERROR | Erro inesperado |
servers:
  - url: http://localhost:3000
...
paths:
...
  /produtos-quimicos/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Produtos químicos]
      summary: Busca produto por id
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Produto' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
    put:
      tags: [Produtos químicos]
      summary: Atualiza um produto (substituição completa)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProdutoInput' }
      responses:
        '200':
          description: Produto atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Produto' }
        '400': { $ref: '#/components/responses/Invalido' }
        '404': { $ref: '#/components/responses/NaoEncontrado' }
...
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
  responses:
    NaoEncontrado:
      description: Recurso não encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
          example: { erro: { codigo: NOT_FOUND, mensagem: Carga química não encontrada } }
    Invalido:
      description: Dados de entrada inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
...
  schemas:
...
    ClasseRisco:
      type: string
      description: Classe/subclasse de risco da ONU
      enum: ['1', '2.1', '2.2', '2.3', '3', '4.1', '4.2', '4.3', '5.1', '5.2', '6.1', '6.2', '7', '8', '9']
    ProdutoInput:
      type: object
      required: [nome, numeroOnu, classeRisco]
      properties:
        nome: { type: string, maxLength: 150 }
        descricao: { type: string, nullable: true }
        numeroOnu: { type: string, example: '1170', description: '4 dígitos, aceita prefixo UN' }
        classeRisco: { $ref: '#/components/schemas/ClasseRisco' }
        grupoCompatibilidade:
          type: string
          nullable: true
          enum: [A, B, C, D, E, F, G, H, J, K, L, N, S]
        status: { type: string, enum: [ATIVO, INATIVO], default: ATIVO }
    Produto:
      allOf:
        - $ref: '#/components/schemas/ProdutoInput'
        - type: object
          properties:
            id: { type: string, format: uuid }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
...
    Erro:
      type: object
      properties:
        erro:
          type: object
          properties:
            codigo: { type: string }
            mensagem: { type: string }
            detalhes: {}